mdbx: simplify work with flags (#12838)

+++ erigon-lib/diagnostics/client.go
@@ -94,7 +94,7 @@ func createDb(ctx context.Context, dbDir string) (db kv.RwDB, err error) {
 		WithTableCfg(func(defaultBuckets kv.TableCfg) kv.TableCfg { return kv.DiagnosticsTablesCfg }).
 		GrowthStep(4 * datasize.MB).
 		MapSize(16 * datasize.GB).
-		PageSize(uint64(4 * datasize.KB)).
+		PageSize(4 * datasize.KB).
 		RoTxsLimiter(semaphore.NewWeighted(9_000)).
 		Path(dbDir).
 		Open(ctx)
